add token revocation logic, refactor into module
