+++ lib/index.js
@@ -12,7 +12,8 @@ var async = require('async')
     , zmq = require('zmq')
     ;
 
-exports.webtask = require('./webtask.js');
+exports.webtask = require('./webtask');
+exports.revocation = require('./revocation');
 
 var last_resort_logger = bunyan.createLogger({ name: 'auth0-coreos-tools-last-resort-logging' });
