+++ lib/webtask.js
@@ -3,9 +3,8 @@ var jws = require('jws')
     , crypto = require('crypto')
     , async = require('async')
     , request = require('request')
-    , aws = require('aws-sdk')
-    , LRU = require('lru-cache')
-    , dynamo
+    , revocation = require('./revocation')
+    , revocation_apis
     ;
 
 var allowed_request_claims = {
@@ -386,30 +385,7 @@ exports.ensure_authorized = function (admin, ctx) {
     };
 };
 
-var lru; // LRU cache of token revocation information
-
-exports.clear_revocation_lru = function (jti) {
-    if (lru) {
-        lru.del(jti);
-    }
-};
-
 exports.ensure_authenticated = function(ctx) {
-
-    if (!lru) {
-        lru = LRU({
-            max: +ctx.config.max_revocation_cache_size.current,
-            maxAge: 1000 * +ctx.config.max_revocation_cache_age.current
-        });
-        ctx.config.max_revocation_cache_size.on('modified', function (new_max) {
-            lru.max = +new_max;
-            ctx.logger.info({ max: +new_max }, 'setting new maximum size of the revocation check LRU cache');
-        });
-        ctx.config.max_revocation_cache_age.on('modified', function (new_age) {
-            lru._maxAge = 1000 * +new_age;
-            ctx.logger.info({ max_age: +new_age }, 'setting new maximum age in seconds of the revocation check LRU cache');
-        });
-    }
 
     // ctx must have logger and config attached to it
     // Let pass requests with:
@@ -602,7 +578,11 @@ exports.ensure_authenticated = function(ctx) {
                         return callback();
                     }
 
-                    check_revocation(jwt, function (error, valid) {
+                    if (!revocation_apis) {
+                        revocation_apis = revocation.init(ctx);
+                    }
+
+                    revocation_apis.check_revocation(req.run_id, jwt, function (error, valid) {
                         if (error) {
                             msg = { code: 502, message: 'revocation check failed', req_id: req.run_id };
                             ctx.logger.error({
@@ -724,95 +704,6 @@ exports.ensure_authenticated = function(ctx) {
                 else {
                     return next();
                 }
-            });
-        }
-
-        function check_revocation(jwt, callback) {
-            // Check revocation status of the JWT token and its issuance chain
-            var ca = [];
-            if (jwt.payload.ca)
-                jwt.payload.ca.forEach(function (jti) { ca.push(jti); });
-            if (jwt.payload.jti)
-                ca.push(jwt.payload.jti);
-            var valid = true;
-            async.each(ca, function (jti, callback) {
-                if (ctx.config.revocation_whitelist.list[jti]) {
-                    // Whitelisted jti
-                    ctx.logger.info({ 
-                        req_id: req.run_id,
-                        jti: jti
-                    }, 'revocation check whitelist match');
-                    return callback();
-                }
-                var jti_valid = lru.get(jti);
-                if (jti_valid === undefined) {
-                    // Revocation info not in cache, check in Dynamo
-
-                    if (!dynamo) {
-                        // Create DynamoDB client 
-                        aws.config.accessKeyId = ctx.config.aws_access_key_id.current;
-                        aws.config.secretAccessKey = ctx.config.aws_secret_access_key.current;
-                        aws.config.region = ctx.config.aws_revocation_region.current;
-                        aws.config.sslEnabled = true;
-                        dynamo = new aws.DynamoDB({ params: { TableName: ctx.config.aws_revocation_table.current }});
-                    }
-
-                    ctx.logger.info({ 
-                        req_id: req.run_id,
-                        jti: jti, 
-                    }, 'starting revocation check in DB');
-
-                    var start = Date.now();
-                    dynamo.query({
-                        KeyConditions: {
-                            jti: { 
-                                ComparisonOperator: 'EQ',
-                                AttributeValueList: [ { S: jti } ]
-                            }
-                        }
-                    }, function (err, data) {
-                        if (err) {
-                            ctx.logger.warn({ 
-                                req_id: req.run_id,
-                                jti: jti, 
-                                dynamo_req_id: this && this.requestId,
-                                revocation_check_latency: Date.now() - start,
-                                error: err.message
-                            }, 'error checking revocation in DB');
-                            return callback(err);
-                        }
-
-                        if (data.Count > 0)
-                            valid = jti_valid = false;
-                        else
-                            jti_valid = true;
-
-                        lru.set(jti, jti_valid);
-
-                        ctx.logger.info({ 
-                            req_id: req.run_id,
-                            jti: jti, 
-                            valid: jti_valid,
-                            dynamo_req_id: this && this.requestId,
-                            revocation_check_latency: Date.now() - start 
-                        }, 'completed revocation check in DB');
-
-                        return callback();
-                    });
-                }
-                else {
-                    // Use revocation info from cache
-                    ctx.logger.info({ 
-                        req_id: req.run_id,
-                        jti: jti, 
-                        valid: jti_valid 
-                    }, 'revocation check cache result');
-                    if (!jti_valid)
-                        valid = false;
-                    return callback();
-                }
-            }, function (error) {
-                callback(error, valid);
             });
         }
     };
